backend/app/services/ai_service/pyproject.toml
[project]
name = "ai_service"
version = "0.1.0"
description = "AI microservice for generation, OCR, ASR, and embeddings"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "shared",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "opentelemetry-instrumentation-requests>=0.46b0",
  "opentelemetry-instrumentation-logging>=0.46b0",
  "langsmith",
  "pymongo>=4.6.0",
  "qdrant-client>=1.7.4",
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langgraph>=0.2.0",
  "sentence-transformers>=2.6.1",
  "cryptography>=42.0.5",
  "cachetools>=5.3.3",
  "ollama>=0.1.9",
  "python-jose>=3.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_service"]
